api/requirements.txt
annotated-types==0.7.0
blinker==1.8.2
click==8.1.7
Flask==3.0.3
gunicorn==22.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
packaging==24.1
psycopg2-binary==2.9.9
pydantic==2.7.4
pydantic_core==2.18.4
redispy==3.0.0
typing_extensions==4.12.2
Werkzeug==3.0.3